split coreml options in isolate file
