+++ onnxruntime/core/providers/coreml/coreml_execution_provider.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "core/common/inlined_containers.h"
+#include "core/providers/coreml/coreml_options.h"
 #include "core/framework/execution_provider.h"
 #include "core/framework/model_metadef_id_generator.h"
 
@@ -11,28 +11,6 @@ namespace onnxruntime {
 namespace coreml {
 class Model;
 }
-
-class CoreMLOptions {
- private:
-  bool require_static_shape_{false};
-  bool create_mlprogram_{false};
-  bool enable_on_subgraph_{false};
-  uint32_t compute_units_{0};
-
- public:
-  explicit CoreMLOptions(uint32_t coreml_flags);
-
-  CoreMLOptions(const ProviderOptions& options) {
-    ValidateAndParseProviderOption(options);
-  }
-  bool RequireStaticShape() const { return require_static_shape_; }
-  bool CreateMLProgram() const { return create_mlprogram_; }
-  bool EnableOnSubgraph() const { return enable_on_subgraph_; }
-  uint32_t ComputeUnits(uint32_t specific_flag = 0xffffffff) const { return compute_units_ & specific_flag; }
-
- private:
-  void ValidateAndParseProviderOption(const ProviderOptions& options);
-};
 
 class CoreMLExecutionProvider : public IExecutionProvider {
  public:
